Support: add local overloads for type-conversion
The imported values are imported as `Int32` rather than `UInt32` which
makes general usage of the values more difficult.  Provide a local
representation which uses the `UInt32` type.

+++ Sources/SwiftWin32/Views and Controls/Label.swift
@@ -9,7 +9,7 @@ import WinSDK
 
 public class Label: Control {
   private static let `class`: WindowClass = WindowClass(named: WC_STATIC)
-  private static let style: WindowStyle = (base: DWORD(WS_TABSTOP), extended: 0)
+  private static let style: WindowStyle = (base: WS_TABSTOP, extended: 0)
 
   public override var font: Font! {
     get { return super.font }
